Experimenting fastest path for straight/diagonal path to end point

+++ AlgorithmSimulator/src/Data/Vertex.java
@@ -4,12 +4,12 @@ import java.util.ArrayList;
 
 public class Vertex implements Comparable<Vertex>
 {
-    public final int x;
-    public final int y;
+    public final float x;
+    public final float y;
     public ArrayList<Vertex> adjacencies = new ArrayList<Vertex>();
     public double minDistance = Double.POSITIVE_INFINITY;
     public Vertex previous;
-    public Vertex(int x, int y) { this.x = x;this.y= y; }
+    public Vertex(float x, float y) { this.x = x;this.y= y; }
     public String toString() 
     { 
     	return "("+x+","+y+")";
